minor improvements
Added fault tolerance for graphing data when GUI did not receive data
from all the motes.
Changed 'hold on' to 'hold all' when plotting raw data values in order
to have the data from different motes be plotted in different colors (up
to 7).
Added code to automatically center the GUI on the screen when it is
created in order to accommodate different screen sizes